+++ lib/hackpad/cli/padlist.rb
@@ -27,12 +27,13 @@ module Hackpad
         all
       end
 
-      def check_list
+      def check_list(refresh=false)
         all = []
         list = Api.list
         list.each do |a|
           pad = Pad.new a
           if !pad.is_cached?
+            pad.load 'txt', refresh
             all << OpenStruct.new( id: a, title: pad.title )
           end
         end