+++ spec/lib/hackpad/cli/api_spec.rb
@@ -1,0 +1,15 @@
+# encoding: utf-8
+
+require 'spec_helper'
+require "hackpad/cli/api"
+
+describe Hackpad::Cli::Api do
+
+  pending ".prepare"
+  pending ".search"
+  pending ".list"
+  pending ".read_options"
+  pending ".read"
+  pending ".get"
+
+end